refactor: باقی مونده از کامیت قبلی مربوط به اصلاحیه و کامل کردن مدل مربوط به تخفیف و اوردر
